Added an initial implementation of PhysicalFolder, for hosting an entire folder and contents. Switched to using it, which required putting the subfolder's index.html actually in a subfolder and actually calling it index.html. Doh. Changed a lot of const char* and char* to dynamic_strings. Made HTTPResponse::SendTo() virtual, so that GetFromPath could return a HTTPResponse instead of sending it itself. This allows PhysicalFolder to call the VirtualFolder::GetFromPath first without it sending a 404 response if it doesn't find anything. Added a version of HTTPResponseFile that takes an already opened FileHandle. Added == operator, SetLen and Normalize to dynamic_string. #pragma once'd the settings header. Added 301 Moved (currently unused) and 403 (completely) Forbidden responses. Finally changed HTTPResponseFile over to using its Headers array, and moved some common headers to the base HTTPResponse class. Squished a type warning in settings.cpp. New win32 .exe.

+++ Src/settings.cpp
@@ -38,8 +38,8 @@ void Settings::save(string filePath){
 }
 
 std::string Settings::strToLower(std::string str){
-	const int length = str.length();
-		for(int i=0; i!=length; ++i){
+	const size_t length = str.length();
+		for(size_t i=0; i!=length; ++i){
 			str[i] = std::tolower(str[i]);
 		}
 	return str;
